python/branchforge_core/pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "branchforge-core"
version = "0.1.0"
description = "BranchForge core generator (heatmap -> branching networks -> exports)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "shapely>=2.0",
  "pydantic>=2.6",
  "networkx>=3.2",
  "Pillow>=10.0",
  "matplotlib>=3.8",
  "reportlab>=4.0",
  "ezdxf>=1.3",
]

[project.optional-dependencies]
cad = ["cadquery>=2.4"]

[tool.hatch.build.targets.wheel]
packages = ["branchforge_core"]
